requirements.txt
google-api-python-client
google-auth-oauthlib
python-dotenv
playwright
jinja2
# winrt-Windows.System.UserProfile  # Windows only — uncomment on Windows
